requirements.txt
# Core LangChain packages
langchain>=0.3.0
langchain-core>=0.3.0
langchain-experimental>=0.0.45
langchain-groq>=0.1.0  # ✅ Use Groq instead of OpenAI

# Environment variable support
python-dotenv>=1.0.0

# Graph visualization
pyvis>=0.3.2

# Web UI
streamlit>=1.32.0